move Settings to user folder and add active class to Profile link menu only if profile is self
